Frontend/Sid-OS/Sid-OS-new/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-docx>=1.1.0
pydantic>=2.6.0

